phase-05/audit-service/pyproject.toml
[project]
name = "audit-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.13.0",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "kafka-python>=2.3.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]
